Starting deprecation drop of PstFromFlopyModel
+ moved method to legacy -- and changed refs in tests
++ also cleaned up some exes and bins -- do we still need to package these in repo?
++ updated toml and coverage config
