--- .coveragerc
@@ -1,12 +1,0 @@
-[run]
-source = pyemu
-parallel = True
-relative_files = True
-omit =
-    */python?.?/*
-    */lib-python/?.?/*.py
-    */lib_pypy/_*.py
-    */site-packages/ordereddict.py
-    */site-packages/nose/*
-    */unittest2/*
-    *__init__*
