+++ pyemu/legacy/__init__.py
@@ -1,0 +1,4 @@
+"""This module contains legacy, deprecated classes and methods,
+including PstFromFlopyModel -- superseeded by `pyemu.utils.PstFrom()`"""
+
+from .pstfromflopy import *
